add_get_key_request_support Move the download keys promise management into a dedicated class.
